Refactor: Move Response handling code to its own module.
The main response handling pipeline still stays under MessageHandler for now,
but I may move it under Response later. Should I do that, I'd also like to
look at making a promise-driven after-events. I might also make
ResponseHandler one of the things that is plugged into MessageHandler if I
do that.

For now, the Response handling module is considered for all intents and
purposes to be an implementation detail of MessageHandler.

I also added 13 more tests for the Response handling code, since it was so
easy to write them. That's 13 more guarantees of quality in every Tennu
release, and that makes me happy.