Changed: Updated irc-socket to > v.3.0.2 + fallout
Also:
Added: "self" plugin. Replaces NicknameTracker object.
Changed: Updated after-events to > v.1.0.1.

The after-events change was because I was wrapping a promise
chain in a setImmediate inside that module, which was making
everything wait two turn cycles instead of a single turn cycle.
This was making some of my more brittle tests fail.

The NickNameTracker object was replaced with a plugin for two
reasons.

1) For one, I want *everything* to be a plugin at some
point. While I originally liked the hard-coded object thing,
there really isn't a point to it. I'd also like to invert the
control of Tennu owning an instance of Tennu-Plugins to
Tennu-Plugins creating the Client object, and then have plugins
add methods to it.

2) The second reason was that I had to rewrite it anyways.
The invariant that the nickname will be the one in the config
value by default was violated because there's now a nickname
array. If I had to rewrite a lot of it, there was no point
in not also making it a plugin at the same time.

And now for the biggy. I changed irc-socket in a major way
to handle startup more correctly. Now instead of just assuming
it is always correct, it actually checks what the server says.
It also has a promise for when `001` happens that returns
the bot's nickname and any CAPs assigned.

The CAPs are important because the next commit is going to make
it so that we'll try to get multi-prefix when we connect.

Some other fallout of updating after-events is that a `done`
become unnecessary (and actually wrong) in the message
handler tests.

Some other fallout of updating irc-socket is that the startup
plugin now uses the irc-socket's startupPromise before doing
its business.