Multiple: [Breaking] Only one command handler
Internal: Refactored PluginCommands. Removed Response handling logic.
Changed: [Breaking] Only allow one command handler per command.

This has been a long time coming. I've been thinking about doing it
for years. By making it so that only one handler is allowed per
command, we simplify the logistics of commands. There was never
really a good reason to allow multiple command handlers other than
it was easy to allow it. But allowing it also allows bad bot
design, and I'd rather ensure that good bot design happens.

Because there's only a maximum of one handler per command, there's
also no need to duplicate the Response handling code in the
Commands plugin. This means that the code only exists in one place,
so the prior issue of fixing the logic in one place and not the
other is now completely gone. But of course, your command handlers
can still return responses.

I made it so that `once` and `off` cannot be used. This is mostly
out of laziness.