Changed: Improve Commands plugin logging
Specifically, all log statements say which command they are for,
and when a command handler is found, it'll now say which plugin
it is from.

Also added a logging statement saying from which plugin a
command is added from when it is being added. I'd make it
info, but that's already devoted to `<-` and `->` and there's
nothing next to it other than debug (which should be disabled
outside of testing) and notice, so notice is what I went with.